fix(Dropdown): unify components in one file
