--- src/components/DropdownHeader/DropdownHeader.scss
@@ -1,20 +1,0 @@
-@import '../../styles/colors';
-@import '../../styles/typography';
-
-.zui-dropdown-header-container {
-  padding: 0.5rem 1rem 1rem 1rem;
-
-  border-bottom: 1px solid #463258;
-
-  user-select: none;
-}
-
-.zui-dropdown-header-title {
-  font-weight: #{$font-weight-bold};
-  font-size: #{$size-default};
-
-  color: #{$text-primary};
-  outline: none;
-
-  user-select: none;
-}
