+++ src/components/DropdownMenu/DropdownMenu.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { cleanup, render, screen } from '@testing-library/react';
 
-import { DropdownMenu, DropdownItem } from '.';
+import * as Dropdown from '.';
 import {
   DropdownMenuRootContentProps,
   DropdownMenuTriggerProps,
@@ -10,13 +10,29 @@ import {
 } from '@radix-ui/react-dropdown-menu';
 
 const mockOnSelect = jest.fn();
-const MOCK_ITEMS: DropdownItem[] = [
+const MOCK_ITEMS: Dropdown.DropdownItem[] = [
   { id: 'dropdown_item_id_1', label: 'Dropdown Item 1', className: 'dropdown-item-1', onSelect: mockOnSelect },
   { id: 'dropdown_item_id_2', label: 'Dropdown Item 2', className: 'dropdown-item-2', onSelect: mockOnSelect },
   { id: 'dropdown_item_id_3', label: 'Dropdown Item 3', className: 'dropdown-item-3', onSelect: mockOnSelect }
 ];
 
-const DEFAULT_PROPS = { items: MOCK_ITEMS };
+const mockText = ['Mock Text', 'Second Mock Text'];
+const mockClassName = 'Mock-ClassName';
+const MockComponent = ({ text }: { text: string }) => {
+  return (
+    <div id="mockComponent">
+      <p>{text}</p>
+    </div>
+  );
+};
+
+const DEFAULT_PROPS_MENU = { items: MOCK_ITEMS };
+
+const DEFAULT_PROPS_HEADER: Dropdown.HeaderProps = {
+  item: '',
+  className: undefined,
+  titleClassName: undefined
+};
 
 /* Mock Radix DropdownMenu primitives so we can skip testing them */
 
@@ -59,7 +75,7 @@ afterEach(() => {
 
 describe('<DropdownMenu />', () => {
   test('should render Radix primitives with correct nesting', () => {
-    const { container } = render(<DropdownMenu {...DEFAULT_PROPS} />);
+    const { container } = render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} />);
 
     const root = container.firstChild;
     const trigger = screen.getByTestId('trigger');
@@ -78,7 +94,7 @@ describe('<DropdownMenu />', () => {
 
   describe('class names', () => {
     test('should apply className to trigger', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} className="mock-class" />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} className="mock-class" />);
 
       expect(screen.getByTestId('trigger')).toHaveClass('mock-class');
     });
@@ -86,13 +102,13 @@ describe('<DropdownMenu />', () => {
 
   describe('menu items', () => {
     test('should apply classNames to menu items', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} className="mock-class" />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} className="mock-class" />);
 
       MOCK_ITEMS.forEach(item => expect(screen.getByText(item.label as string)).toHaveClass(item.className));
     });
 
     test('should render correct menu item labels', () => {
-      const { container } = render(<DropdownMenu {...DEFAULT_PROPS} />);
+      const { container } = render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} />);
 
       const items = container.getElementsByClassName('mock-item');
       expect(items.length).toBe(MOCK_ITEMS.length);
@@ -100,7 +116,7 @@ describe('<DropdownMenu />', () => {
     });
 
     test('should make menu items clickable', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} />);
 
       mockRadixItem.mock.lastCall[0].onSelect();
       expect(mockOnSelect).toBeCalledTimes(1);
@@ -109,13 +125,13 @@ describe('<DropdownMenu />', () => {
 
   describe('dropdown trigger', () => {
     test('should render `...` as default trigger', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} />);
 
       expect(screen.getByTestId('trigger').textContent).toBe('...');
     });
 
     test('should render string trigger', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} trigger="mock-trigger-string" />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} trigger="mock-trigger-string" />);
 
       const triggerText = screen.getByText('mock-trigger-string');
       expect(triggerText.nodeName).toBe('SPAN');
@@ -123,7 +139,7 @@ describe('<DropdownMenu />', () => {
     });
 
     test('should render component trigger', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} trigger={<aside>mock trigger component</aside>} />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} trigger={<aside>mock trigger component</aside>} />);
 
       const triggerText = screen.getByText('mock trigger component');
       expect(triggerText.nodeName).toBe('ASIDE');
@@ -133,15 +149,74 @@ describe('<DropdownMenu />', () => {
 
   describe('@radix-ui prop forwarding', () => {
     test('should pass content props to Radix Content component', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} alignMenu="start" side="top" />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} alignMenu="start" side="top" />);
 
       expect(mockRadixContent).toBeCalledWith(expect.objectContaining({ align: 'start', side: 'top' }));
     });
 
     test('should pass root props to Radix Root component', () => {
-      render(<DropdownMenu {...DEFAULT_PROPS} open={true} defaultOpen={true} />);
+      render(<Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} open={true} defaultOpen={true} />);
 
       expect(mockRadixRoot).toBeCalledWith(expect.objectContaining({ open: true, defaultOpen: true }));
     });
+  });
+
+  describe('menu header component', () => {
+    test('should render Header component', () => {
+      const { container } = render(
+        <Dropdown.DropdownMenu {...DEFAULT_PROPS_MENU} open={true} defaultOpen={true}>
+          <Dropdown.Header {...DEFAULT_PROPS_HEADER} className={mockClassName} item={mockText[0]} />
+        </Dropdown.DropdownMenu>
+      );
+
+      expect(container.getElementsByClassName(mockClassName).length).toBe(1);
+      expect(screen.getByText(mockText[0])).toBeInTheDocument();
+    });
+  });
+});
+
+const renderHeader = (props?: Dropdown.HeaderProps) => render(<Dropdown.Header {...DEFAULT_PROPS_HEADER} {...props} />);
+
+describe('<Dropdown.Header />', () => {
+  test('should recognize a string and render it from props in a <span> tag', () => {
+    const { container } = renderHeader({ item: mockText[0] });
+    const item = screen.getByText(mockText[0]);
+
+    expect(container.firstChild?.textContent).toBe(mockText[0]);
+    expect(item.tagName).toBe('SPAN');
+  });
+
+  test('should recognize a Component and render it from props', () => {
+    renderHeader({ item: <MockComponent text={mockText[0]} /> });
+
+    expect(screen.getByText(mockText[0])).toBeInTheDocument();
+  });
+
+  test('should recognize a Fragment and render it from props', () => {
+    renderHeader({
+      item: (
+        <>
+          <MockComponent text={mockText[0]} />
+          <MockComponent text={mockText[1]} />
+        </>
+      )
+    });
+
+    mockText.forEach(item => expect(screen.getByText(item)).toBeInTheDocument());
+  });
+
+  test('should apply className from props to tag container', () => {
+    const { container } = renderHeader({ ...DEFAULT_PROPS_HEADER, className: mockClassName });
+
+    expect(container.firstChild).toHaveClass(mockClassName);
+    expect(container.getElementsByClassName(mockClassName).length).toBe(1);
+  });
+
+  test('should recognize a string and apply className from props to <span> tag', () => {
+    const { container } = renderHeader({ ...DEFAULT_PROPS_HEADER, item: mockText[0], titleClassName: mockClassName });
+    const item = screen.getByText(mockText[0]);
+
+    expect(item).toHaveClass(mockClassName);
+    expect(container.getElementsByClassName(mockClassName).length).toBe(1);
   });
 });
