+++ src/components/index.ts
@@ -26,4 +26,3 @@ export * from './TextStack';
 export * from './ToggleGroup';
 export * from './Tooltip';
 export { Wizard } from './Wizard';
-
